plot_items: set all subplots to the same X range.
When plotting multiple channels (of signals and/or annotations), the X
axes should be aligned so that simultaneous events in different
channels are horizontally aligned with each other.

Since plot_items and plot_wfdb don't currently support multi-frequency
signal data, this won't make a difference in common cases; it does,
however, matter if you are plotting *only* annotations (without
signals), or if there are more annotation channels than signal
channels.